rename useCancelOrder to useCancel for consistency
